Wire enro-test and add UkptViewModel unit tests
Add enro-test to the version catalog and :feature:core:client commonTest
dependencies. Write UkptViewModelTest covering load lifecycle, error + retry,
greet action, and dialog result channel (completed vs closed). Move the
domain/infrastructure tests out of UkptViewModelTest.kt into files named
after their class under test.
