+++ feature/core/client/src/commonTest/kotlin/feature/ukpt/client/ui/UkptViewModelTest.kt
@@ -1,173 +1,201 @@
+@file:OptIn(ExperimentalCoroutinesApi::class)
+
 package feature.ukpt.client.ui
 
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import dev.enro.result.NavigationResultChannel
+import dev.enro.test.assertOpened
+import dev.enro.test.putNavigationHandleForViewModel
+import dev.enro.test.runEnroTest
+import dev.enro.test.sendClosedForTest
+import dev.enro.test.sendCompletedForTest
 import dev.isaacudy.udytils.state.AsyncState
-import dev.isaacudy.udytils.state.fromFlow
-import dev.isaacudy.udytils.state.fromSuspending
-import dev.isaacudy.udytils.state.isLoading
-import dev.isaacudy.udytils.state.isSuccess
 import feature.ukpt.Greeting
-import feature.ukpt.client.data.GreetingRepository
-import feature.ukpt.client.domain.FlowOfGreetingHistory
-import feature.ukpt.client.domain.FlowOfGreetingSummaryImpl
-import feature.ukpt.client.domain.FlowOfLatestGreeting
+import feature.ukpt.client.domain.FlowOfGreetingSummary
+import feature.ukpt.client.domain.GetGreeting
 import feature.ukpt.client.domain.GreetingSummary
+import feature.ukpt.client.domain.ResetGreetings
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.toList
-import kotlinx.coroutines.test.runTest
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
+import kotlin.test.AfterTest
+import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertIs
-import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
-class FlowOfGreetingSummaryImplTest {
+class UkptViewModelTest {
+
+    private val testDispatcher = UnconfinedTestDispatcher()
+    private val createdViewModels = mutableListOf<ViewModel>()
+
+    private fun <T : ViewModel> T.track(): T {
+        createdViewModels += this
+        return this
+    }
+
+    @BeforeTest
+    fun setUp() {
+        Dispatchers.setMain(testDispatcher)
+    }
+
+    @AfterTest
+    fun tearDown() {
+        // Cancel each ViewModel's scope before resetting Main: a collector left subscribed
+        // to pendingResults dispatches to a Main dispatcher that no longer exists on targets
+        // without a default one, failing an unrelated test.
+        createdViewModels.forEach { it.viewModelScope.cancel() }
+        createdViewModels.clear()
+        NavigationResultChannel.pendingResults.value = emptyMap()
+        Dispatchers.resetMain()
+    }
 
     @Test
-    fun combinesLatestGreetingAndHistory() = runTest {
-        val impl = FlowOfGreetingSummaryImpl(
-            flowOfLatestGreeting = FlowOfLatestGreeting { flowOf(Greeting(text = "Hello")) },
-            flowOfGreetingHistory = FlowOfGreetingHistory {
-                flowOf(listOf(Greeting(text = "Hello"), Greeting(text = "Hi")))
+    fun loadLifecycleReachesSuccessOnEmission() = runEnroTest {
+        val summaryFlow = MutableStateFlow(
+            GreetingSummary(
+                latestGreeting = Greeting(text = "Hello"),
+                greetingHistory = listOf(Greeting(text = "Hello")),
+            )
+        )
+        putNavigationHandleForViewModel<UkptViewModel, UkptDestination>(UkptDestination)
+
+        val vm = UkptViewModel(
+            flowOfGreetingSummary = FlowOfGreetingSummary { summaryFlow },
+            getGreeting = GetGreeting { "Hello" },
+            resetGreetings = ResetGreetings { },
+        ).track()
+
+        val state = vm.state.value
+        assertIs<AsyncState.Success<GreetingSummary>>(state.greetingSummary)
+        assertEquals("Hello", state.greetingSummary.data.latestGreeting?.text)
+        assertEquals(1, state.greetingSummary.data.greetingHistory.size)
+    }
+
+    @Test
+    fun loadErrorLandsInAsyncStateError() = runEnroTest {
+        var shouldFail = true
+        val summaryFlow = MutableStateFlow(
+            GreetingSummary(
+                latestGreeting = null,
+                greetingHistory = emptyList(),
+            )
+        )
+
+        putNavigationHandleForViewModel<UkptViewModel, UkptDestination>(UkptDestination)
+
+        val vm = UkptViewModel(
+            flowOfGreetingSummary = FlowOfGreetingSummary {
+                if (shouldFail) {
+                    flow { throw IllegalStateException("load failed") }
+                } else {
+                    summaryFlow
+                }
             },
+            getGreeting = GetGreeting { "Hello" },
+            resetGreetings = ResetGreetings { },
+        ).track()
+
+        assertIs<AsyncState.Error<GreetingSummary>>(vm.state.value.greetingSummary)
+
+        shouldFail = false
+        vm.onRetryClicked()
+
+        assertIs<AsyncState.Success<GreetingSummary>>(vm.state.value.greetingSummary)
+    }
+
+    @Test
+    fun greetActionDrivesStateToSuccess() = runEnroTest {
+        val summaryFlow = MutableStateFlow(
+            GreetingSummary(latestGreeting = null, greetingHistory = emptyList())
+        )
+        var greetCalled = false
+
+        putNavigationHandleForViewModel<UkptViewModel, UkptDestination>(UkptDestination)
+
+        val vm = UkptViewModel(
+            flowOfGreetingSummary = FlowOfGreetingSummary { summaryFlow },
+            getGreeting = GetGreeting {
+                greetCalled = true
+                "Hello"
+            },
+            resetGreetings = ResetGreetings { },
+        ).track()
+
+        vm.onGreetClicked()
+
+        assertTrue(greetCalled)
+        assertIs<AsyncState.Success<Unit>>(vm.state.value.greetAction)
+    }
+
+    @Test
+    fun greetActionErrorLandsInAsyncStateError() = runEnroTest {
+        val summaryFlow = MutableStateFlow(
+            GreetingSummary(latestGreeting = null, greetingHistory = emptyList())
         )
 
-        val result = impl().first()
-        assertEquals(Greeting(text = "Hello"), result.latestGreeting)
-        assertEquals(2, result.greetingHistory.size)
+        putNavigationHandleForViewModel<UkptViewModel, UkptDestination>(UkptDestination)
+
+        val vm = UkptViewModel(
+            flowOfGreetingSummary = FlowOfGreetingSummary { summaryFlow },
+            getGreeting = GetGreeting { throw IllegalStateException("greet failed") },
+            resetGreetings = ResetGreetings { },
+        ).track()
+
+        vm.onGreetClicked()
+
+        assertIs<AsyncState.Error<Unit>>(vm.state.value.greetAction)
     }
 
     @Test
-    fun handlesNullLatestGreeting() = runTest {
-        val impl = FlowOfGreetingSummaryImpl(
-            flowOfLatestGreeting = FlowOfLatestGreeting { flowOf(null) },
-            flowOfGreetingHistory = FlowOfGreetingHistory { flowOf(emptyList()) },
+    fun resetViaDialogCompletionTriggersResetGreetings() = runEnroTest {
+        val summaryFlow = MutableStateFlow(
+            GreetingSummary(latestGreeting = null, greetingHistory = emptyList())
         )
+        var resetCalled = false
 
-        val result = impl().first()
-        assertNull(result.latestGreeting)
-        assertTrue(result.greetingHistory.isEmpty())
+        val handle = putNavigationHandleForViewModel<UkptViewModel, UkptDestination>(UkptDestination)
+
+        val vm = UkptViewModel(
+            flowOfGreetingSummary = FlowOfGreetingSummary { summaryFlow },
+            getGreeting = GetGreeting { "Hello" },
+            resetGreetings = ResetGreetings { resetCalled = true },
+        ).track()
+
+        vm.onResetRequested()
+        val child = handle.assertOpened<ConfirmResetDestination>()
+        child.sendCompletedForTest()
+
+        assertTrue(resetCalled)
     }
 
     @Test
-    fun updatesWhenSourceChanges() = runTest {
-        val latest = MutableStateFlow<Greeting?>(null)
-        val history = MutableStateFlow<List<Greeting>>(emptyList())
-        val impl = FlowOfGreetingSummaryImpl(
-            flowOfLatestGreeting = FlowOfLatestGreeting { latest },
-            flowOfGreetingHistory = FlowOfGreetingHistory { history },
+    fun resetDialogClosedDoesNotTriggerResetGreetings() = runEnroTest {
+        val summaryFlow = MutableStateFlow(
+            GreetingSummary(latestGreeting = null, greetingHistory = emptyList())
         )
+        var resetCalled = false
 
-        val result1 = impl().first()
-        assertNull(result1.latestGreeting)
-        assertTrue(result1.greetingHistory.isEmpty())
+        val handle = putNavigationHandleForViewModel<UkptViewModel, UkptDestination>(UkptDestination)
 
-        latest.value = Greeting(text = "Hello")
-        history.value = listOf(Greeting(text = "Hello"))
+        val vm = UkptViewModel(
+            flowOfGreetingSummary = FlowOfGreetingSummary { summaryFlow },
+            getGreeting = GetGreeting { "Hello" },
+            resetGreetings = ResetGreetings { resetCalled = true },
+        ).track()
 
-        val result2 = impl().first()
-        assertEquals(Greeting(text = "Hello"), result2.latestGreeting)
-        assertEquals(1, result2.greetingHistory.size)
-    }
-}
+        vm.onResetRequested()
+        val child = handle.assertOpened<ConfirmResetDestination>()
+        child.sendClosedForTest()
 
-class GreetingRepositoryTest {
-
-    @Test
-    fun getGreetingAppendsToHistory() = runTest {
-        val repo = GreetingRepository()
-        repo.getGreeting()
-        val history = repo.flowOfGreetingHistory().first()
-        assertEquals(1, history.size)
-        assertEquals("Hello", history.first().text)
-    }
-
-    @Test
-    fun getGreetingUpdatesLatest() = runTest {
-        val repo = GreetingRepository()
-        assertNull(repo.flowOfLatestGreeting().first())
-
-        repo.getGreeting()
-        assertEquals(Greeting(text = "Hello"), repo.flowOfLatestGreeting().first())
-    }
-
-    @Test
-    fun resetGreetingsClearsList() = runTest {
-        val repo = GreetingRepository()
-        repo.getGreeting()
-        repo.getGreeting()
-        assertEquals(2, repo.flowOfGreetingHistory().first().size)
-
-        repo.resetGreetings()
-        assertTrue(repo.flowOfGreetingHistory().first().isEmpty())
-        assertNull(repo.flowOfLatestGreeting().first())
-    }
-
-    @Test
-    fun multipleGreetsAccumulate() = runTest {
-        val repo = GreetingRepository()
-        repo.getGreeting()
-        repo.getGreeting()
-        repo.getGreeting()
-
-        val history = repo.flowOfGreetingHistory().first()
-        assertEquals(3, history.size)
-        assertEquals(Greeting(text = "Hello"), repo.flowOfLatestGreeting().first())
-    }
-}
-
-class AsyncStateLoadTest {
-
-    @Test
-    fun fromFlowProducesLoadingThenSuccess() = runTest {
-        val summary = GreetingSummary(
-            latestGreeting = Greeting(text = "Hello"),
-            greetingHistory = listOf(Greeting(text = "Hello")),
-        )
-        val states = AsyncState.fromFlow(flowOf(summary)).toList()
-
-        assertTrue(states[0].isLoading())
-        assertIs<AsyncState.Success<GreetingSummary>>(states[1])
-        assertEquals(summary, (states[1] as AsyncState.Success).data)
-    }
-
-    @Test
-    fun fromFlowCapturesError() = runTest {
-        val states = AsyncState.fromFlow(
-            flow<GreetingSummary> { throw IllegalStateException("connection failed") }
-        ).toList()
-
-        assertTrue(states[0].isLoading())
-        assertIs<AsyncState.Error<GreetingSummary>>(states[1])
-    }
-
-    @Test
-    fun fromSuspendingCapturesError() = runTest {
-        val states = AsyncState.fromSuspending<Unit> {
-            throw IllegalStateException("network error")
-        }.toList()
-
-        assertTrue(states[0].isLoading())
-        assertIs<AsyncState.Error<Unit>>(states[1])
-    }
-
-    @Test
-    fun fromSuspendingRetryViaRelaunch() = runTest {
-        var callCount = 0
-        val run = {
-            AsyncState.fromSuspending<Unit> {
-                callCount++
-                if (callCount == 1) throw IllegalStateException("first attempt fails")
-            }
-        }
-
-        val first = run().toList()
-        assertIs<AsyncState.Error<Unit>>(first.last())
-
-        val second = run().toList()
-        assertTrue(second.last().isSuccess())
-        assertEquals(2, callCount)
+        // Absence is the assertion: sendClosedForTest must not invoke resetGreetings.
+        assertTrue(!resetCalled)
     }
 }
